+++ custom_pos_receipt/controllers/__init__.py
@@ -1,1 +1,1 @@
-from . import optometrist_assessment
+from . import main
