+++ custom_pos_receipt/models/__init__.py
@@ -1,2 +1,2 @@
-from . import pos_order
-from . import OptometristAssessment, OptometristObservation
+from . import optometrist_observation, pos_order
+from . import optometrist_assessment
