+++ custom_pos_receipt/models/optometrist_assessment.py
